12/21_Ajax

+++ app/Http/Controllers/PostController.php
@@ -28,7 +28,7 @@ class PostController extends Controller
   }
 
   public function show(Request $request) {
-    $post = Post::find($request->id);
+    $post = Post::findOrFail($request->id);
     $user_id = Auth::id();
     return view('posts.show', compact('post','user_id'));
   }
@@ -53,14 +53,14 @@ class PostController extends Controller
 
   public function edit(Request $request)
   {
-    $post = Post::find($request->id);
+    $post = Post::findOrFail($request->id);
     $this->authorize('view', $post);
     return view('posts.edit', ['post' => $post ]);
   }
 
   public function update(Request $request, Post $post)
   {
-    $post = Post::find($request->id);
+    $post = Post::findOrFail($request->id);
     $rules = [
         'content' => ['required', 'string', 'max:140']
     ];
@@ -72,8 +72,15 @@ class PostController extends Controller
 
   public function destroy(Request $request, Post $post)
   {
-    $post = Post::find($request->id);
-    $post->delete();
-    return redirect('/posts/index');
+    $post = Post::findOrFail($request->id);
+    // $post->delete();
+    if($request->ajax()) {
+      $post->delete();
+      return response()->json(['deleted' => true, 'post' => $post]);
+    }
+     // else {
+    //   $user->delete();
+    //   return response()->setContent(true);
+    // }
   }
 }
